+++ .gitignore
@@ -1,4 +1,4 @@
-/data
+/output
 /docs
 
 # TypeScript Output
