+++ .vscode/settings.json
@@ -2,5 +2,8 @@
     "cSpell.words": [
         "typedoc"
     ],
-    "typescript.tsdk": "node_modules\\typescript\\lib"
+    "typescript.tsdk": "node_modules\\typescript\\lib",
+    "yaml.schemas": {
+        "config/config.schema.json": "config/config.yml"
+    }
 }
