+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "bulletin-issues-transferred",
-  "version": "2.0.0-alpha",
+  "version": "2.0.0-alpha.1",
   "description": "汇总BIT各种网站的通知。",
   "main": "update_notices.js",
   "scripts": {
     "test": "mocha dist/**/*.test.js",
     "build": "tsc",
     "doc": "typedoc",
-    "update": "node dist/src/update_notices.js"
+    "update": "node dist/update_notices.js"
   },
   "author": "Y.D.X.",
   "license": "MIT",
@@ -18,7 +18,8 @@
     "jsdom": "^19.0.0",
     "node-fetch": "^3.1.0",
     "typescript": "^4.5.3",
-    "xml": "^1.0.1"
+    "xml": "^1.0.1",
+    "yaml": "^2.0.0-10"
   },
   "devDependencies": {
     "@types/chai": "^4.3.0",
