+++ src/core/notice.ts
@@ -7,34 +7,9 @@ import { JSDOM } from 'jsdom'
 import chalk from "chalk"
 
 import { parse_date } from '../util/my_date.js'
+import { SourceRaw, SourceInterface, NoticeRaw, NoticeInterface } from "./interfaces.js"
 
 
-
-/**
- * 通知来源（接口）
- * @see `notice_sources.schema.json`
- */
-export interface SourceInterface {
-    name: string
-    full_name: string
-    alt_name: string[]
-    url: string
-    guide: string[]
-
-    /**
-     * 抓取通知
-     */
-    fetch_notice(): Promise<Notice[]>
-}
-
-/**
- * 通知来源（简易格式），可用作构造{@link Source}的参数
- * @see `notice_sources.schema.json`
- */
-export interface SourceRaw {
-    name: string, full_name?: string, alt_name?: string[],
-    url: string, guide?: string[],
-}
 
 /**
  * 通知来源
@@ -125,41 +100,6 @@ export class SourceBySelectors extends Source {
 }
 
 
-
-/**
- * 通知（接口）
- */
-export interface NoticeInterface {
-    link: string
-    title: string
-    date: Date | null
-    source: SourceInterface
-    source_name: string
-
-    /** 转换成容易存储的`NoticeRaw`（其`source`为`string`） */
-    to_raw(): NoticeRaw
-    /** 等同于{@link to_raw} */
-    valueOf(): NoticeRaw
-
-    /**
-     * 转换为对人友好的格式
-     * 
-     * 第一行是来源和标题，第二行是链接，第三行是日期。
-     */
-    to_human_readable_rows(): string[]
-}
-
-/**
- * 通知（简易格式），可用作构造{@link Notice}的参数
- * 
- * `source`既允许{@link Source}，也允许`string`。
- */
-export interface NoticeRaw {
-    link: string
-    title: string
-    date: Date | null
-    source?: SourceInterface | string
-}
 
 /**
  * 通知
