+++ src/core/notices_saver.ts
@@ -5,7 +5,9 @@
 import { readFile, writeFile } from 'fs/promises'
 import chalk from "chalk"
 
-import { NoticeInterface, NoticeRaw, Notice } from "./notice.js"
+import config from './config/config.js'
+import { NoticeRaw, NoticeInterface } from './interfaces.js'
+import { Notice } from "./notice.js"
 import { import_sources } from './sources_importer.js'
 
 
@@ -22,13 +24,13 @@ function json_date_reviver(key: string, value: any) {
 
 
 /**
- * 从`data/notices.json`读取已有通知
+ * 从`config.output.json_path`读取已有通知
  * @param options 选项
  * @param options.ignore_source 是否忽略来源。因为存储来源时只记`name`，读取时可能需要还原为{@link Source}。
  */
 export async function read_json({ ignore_source = false } = {}) {
     try {
-        const json_str = (await readFile('data/notices.json')).toString()
+        const json_str = (await readFile(config.output.json_path)).toString()
         const json: NoticeRaw[] = JSON.parse(json_str, json_date_reviver)
 
         if (ignore_source) {
@@ -50,11 +52,11 @@ export async function read_json({ ignore_source = false } = {}) {
 }
 
 /**
- * 将通知写入`data/notices.json`（不是追加）
+ * 将通知写入`config.output.json_path`（不是追加）
  * @param notices 
  */
 export async function write_json(notices: NoticeInterface[]) {
     const json = JSON.stringify(notices.map(n => n.to_raw()), null, 2)
-    await writeFile('data/notices.json', json)
-    console.log(chalk.green('✓'), '已保存到 data/notices.json。')
+    await writeFile(config.output.json_path, json)
+    console.log(chalk.green('✓'), `已保存到 ${config.output.json_path}。`)
 }
