+++ src/core/sources_importer.ts
@@ -5,23 +5,15 @@
 import { readFile } from 'fs/promises'
 import chalk from "chalk"
 
-import { SourceRaw, SourceInterface } from "./notice.js"
+import config from './config/config.js'
+import { SourceStorageFormat, SourceInterface } from "./interfaces"
 import { SourceBySelectorsRaw, SourceBySelectors } from './notice.js'
-import { parse_special_source } from './sources_special.js'
+import { parse_special_source } from './config/sources_special.js'
 
 
-
-/**
- * 通知来源的存储格式，基本等同于JSON文件
- * @see `notice_sources.schema.json`
- */
-export interface SourceStorageFormat extends SourceRaw {
-    fetch_by: string,
-    selectors?: any
-}
 
 async function _import_sources() {
-    const file = await readFile('config/notice_sources.json')
+    const file = await readFile(config.sources_path)
     const raw_sources: SourceStorageFormat[] = JSON.parse(file.toString()).sources
 
     const sources = raw_sources.map(s => {
@@ -44,7 +36,7 @@ async function _import_sources() {
 let _sources_cache: SourceInterface[] = []
 
 /**
- * 从`config/notice_sources.json`和{@link sources_special}导入通知来源
+ * 从`config.sources_path`和{@link sources_special}导入通知来源
  * @param options 选项 
  * @param options.force 是否强制刷新缓存。因为通知来源基本不变，默认会自动采用缓存。
  */
@@ -54,14 +46,14 @@ export async function import_sources({ force = false } = {}) {
             _sources_cache = await _import_sources()
         } catch (error) {
             if (error.code === 'ENOENT') {
-                console.error(chalk.red('✗ 未找到任何通知来源。您可能需要新建 config/notice_sources.json。'))
+                console.error(chalk.red(`✗ 未找到任何通知来源。您可能需要新建 ${config.sources_path}。`))
             }
             throw error
         }
     }
 
     if (_sources_cache.length === 0) {
-        console.log(chalk.red('✗ 未找到任何通知来源。config/notice_sources.json 可能是空的。'))
+        console.log(chalk.red(`✗ 未找到任何通知来源。${config.sources_path} 可能是空的。`))
     }
 
     return _sources_cache
